packages/common/pyproject.toml
[project]
name = "task-system-common"
version = "0.1.0"
description = "Shared domain and infrastructure code for the task system."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pillow>=11.2.1",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.0",
  "redis>=6.0.0",
  "rq>=2.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/task_system_common"]
